Brings studio/m-a-rail-header into main: pty-output-buffer replay offset
reconciliation (c7a39ef), F-9 Stronghold settings-secret env injection at both
sidecar spawn sites (9711fa0, live-verified: real FAL_KEY in the sidecar environ
with env -u FAL_KEY), and the two dead /iyke/logs filters revived (76530c9,
live-verified since=u64 + level at-or-above-severity).

+++ src-tauri/src/iyke/handlers.rs
@@ -817,8 +817,11 @@ fn require_one_target(
 pub struct LogsQuery {
     #[serde(default)]
     pub level: Option<String>,
+    // u64, not u128: axum's Query uses serde_urlencoded which does NOT support
+    // u128 (every `?since=…` was rejected with "u128 is not supported").
+    // unix-ms fits u64; `recent_logs` casts to u128 at the comparison site.
     #[serde(default)]
-    pub since: Option<u128>,
+    pub since: Option<u64>,
     #[serde(default)]
     pub source: Option<String>,
 }
